refactor: Organize imports and remove unused properties in post components; add Post type definitions
