+++ src/components/postBody.tsx
@@ -1,6 +1,6 @@
 import Markdown from 'markdown-to-jsx'
 import Post_embed from '@/components/ui/postEmbed'
-import { StH2, StH3, StP, A } from './post-strike-through'
+import { A, StH2, StH3, StP } from './post-strike-through'
 
 export default ({ content }: { content: string }) => {
   return (
